Security use Company entity as user provider
