+++ config/packages/security.yaml
@@ -6,7 +6,7 @@ security:
     providers:
         app_user_provider:
             entity:
-                class: App\Entity\User
+                class: App\Entity\Company
                 property: username
     firewalls:
         login:
