+++ src/Entity/User.php
@@ -17,7 +17,7 @@ use Symfony\Component\Validator\Constraints as Assert;
  * @UniqueEntity (fields="username", message="This username is already used")
  * @UniqueEntity (fields="email", message="This email address is already used")
  */
-class User implements UserInterface
+class User
 {
     const USER_COMPANY_ADMIN = 'ROLE_COMPANY_ADMIN';
     const USER_ADMIN = 'ROLE_ADMIN';
@@ -30,53 +30,6 @@ class User implements UserInterface
      */
     private $id;
 
-    /**
-     * @ORM\Column(type="string", length=180, unique=true)
-     * @Groups({"show_user", "list_users", "show_company"})
-     * @Assert\NotBlank (message="Username is missing or empty")
-     * @Assert\Regex(pattern="/^[a-z]+[a-z0-9]+$/i", message="Username must start by a letter and can only consist of alphanumeric characters")
-     * @Assert\Length (
-     *     min = 3,
-     *     max = 30,
-     *     minMessage = "Username must be at least {{ limit }} characters long",
-     *     maxMessage = "Username cannot be longer than {{ limit }} characters",
-     *     )
-     */
-    private $username;
-
-    /**
-     * @ORM\Column(type="json")
-     */
-    private $roles = [];
-
-    /**
-     * @var string $plainPassword the plain password, not stored
-     * @UserPassword()
-     * @Assert\NotBlank(message = "Password is missing or empty")
-     * @Assert\Length(
-     *     max = 255,
-     *     maxMessage="Password cannot be longer than {{limit}} characters"
-     * )
-     * @PasswordRequirements(
-     *      minLength = 8,
-     *      tooShortMessage = "Password must be at least {{ limit }} characters long",
-     *      requireLetters = true,
-     *      missingLettersMessage = "Password must contains at least one letter",
-     *      requireCaseDiff = true,
-     *      requireCaseDiffMessage = "Password must contains lower and upper case letters",
-     *      requireNumbers = true,
-     *      missingNumbersMessage = "Password must contains at least one number",
-     *      requireSpecialCharacter = true,
-     *      missingSpecialCharacterMessage = "Password must contains at least one special character"
-     * )
-     */
-    private $plainPassword;
-
-    /**
-     * @var string The hashed password
-     * @ORM\Column(type="string")
-     */
-    private $password;
 
     /**
      * @ORM\Column(type="string", length=255)
@@ -125,73 +78,6 @@ class User implements UserInterface
         return $this->id;
     }
 
-    /**
-     * A visual identifier that represents this user.
-     *
-     * @see UserInterface
-     */
-    public function getUsername(): string
-    {
-        return (string)$this->username;
-    }
-
-    public function setUsername(string $username): self
-    {
-        $this->username = $username;
-
-        return $this;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getRoles(): array
-    {
-        $roles = $this->roles;
-        // guarantee every user at least has ROLE_USER
-        $roles[] = 'ROLE_USER';
-
-        return array_unique($roles);
-    }
-
-    public function setRoles(array $roles): self
-    {
-        $this->roles = $roles;
-
-        return $this;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getPassword(): string
-    {
-        return (string)$this->password;
-    }
-
-    public function setPassword(string $password): self
-    {
-        $this->password = $password;
-
-        return $this;
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function getSalt()
-    {
-        // not needed when using the "bcrypt" algorithm in security.yaml
-    }
-
-    /**
-     * @see UserInterface
-     */
-    public function eraseCredentials()
-    {
-        // If you store any temporary, sensitive data on the user, clear it here
-        // $this->plainPassword = null;
-    }
 
     public function getEmail(): ?string
     {
@@ -239,21 +125,5 @@ class User implements UserInterface
         $this->company = $company;
 
         return $this;
-    }
-
-    /**
-     * @return string
-     */
-    public function getPlainPassword(): string
-    {
-        return $this->plainPassword;
-    }
-
-    /**
-     * @param string $plainPassword
-     */
-    public function setPlainPassword(string $plainPassword): void
-    {
-        $this->plainPassword = $plainPassword;
     }
 }
